refactor: update imports for `chakra` to `chakra-ui`
Renamed all instances of `chakra` to `chakra-ui` across the codebase for consistency with updated library structure. Ensures clarity, better alignment with library conventions, and eliminates potential confusion in imports.
