+++ apps/nextjs/src/components/layout/navigation/custom-nav/favorites/workspace/workspace-list/list-item/workspace-menu/workspace-menu.tsx
@@ -3,7 +3,7 @@ import { memo, useCallback } from 'react';
 import { Icon } from '@/components/ui/icon';
 import { Menu } from '@/components/ui/menu';
 import { useClickableHoverStyle } from '@/hooks/styles/use-clickable-hover-style';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import { MenuList } from './menu-list';
 
 type Props = {
