+++ apps/nextjs/src/components/pages/my-tasks/components/files/skeleton-files.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 import { Flex, type FlexProps } from '@/components/ui/flex';
 import { Skeleton } from '@/components/ui/skeleton';
 import { Stack } from '@/components/ui/stack';
-import { useBreakpointValue } from '@/lib/chakra';
+import { useBreakpointValue } from '@/lib/chakra-ui';
 
 type Props = FlexProps;
