+++ apps/nextjs/src/components/pages/workspaces/components/overview/overview-timeline/use-timeline-style.ts
@@ -1,5 +1,5 @@
 import { useMemo } from 'react';
-import type { SystemStyleObject } from '@/lib/chakra';
+import type { SystemStyleObject } from '@/lib/chakra-ui';
 
 export const useTimelineStyle = () => {
   const timelineBorderStyle = useMemo(
