+++ apps/nextjs/src/components/ui/editor/tool-bar/format/format.tsx
@@ -4,7 +4,7 @@ import { Icon } from '@/components/ui/icon';
 import { IconButton } from '@/components/ui/icon-button';
 import { Link } from '@/components/ui/link';
 import { Popover } from '@/components/ui/popover';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import {
   useBold,
   useBulletList,
