+++ apps/nextjs/src/components/ui/forms/date-picker-with-input/input.tsx
@@ -2,7 +2,7 @@ import type React from 'react';
 import { memo, useCallback, useState } from 'react';
 import { Input as AtomsInput } from '@/components/ui/input';
 import { PopoverDueDatePicker } from '@/features/task/components/popover-due-date-picker/popover-due-date-picker';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import { formatDueDateInput } from '@/lib/date';
 
 type Props = {
