+++ apps/nextjs/src/components/ui/icon-button/icon-button.tsx
@@ -6,7 +6,7 @@ import type React from 'react';
 import { forwardRef, useMemo } from 'react';
 import { useDisabledStyle } from '@/hooks/styles/use-disabled-style';
 import { useLinkHoverStyle } from '@/hooks/styles/use-link-hover-style';
-import type { SystemStyleObject } from '@/lib/chakra';
+import type { SystemStyleObject } from '@/lib/chakra-ui';
 
 type Props = ChakraIconButtonProps & {
   light?: boolean;
