+++ apps/nextjs/src/components/ui/tooltip/use-tooltip.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useHover } from '@/hooks/use-hover';
 import { useMountedRef } from '@/hooks/use-mounted-ref';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 
 type Props = {
   openDelay?: number;
