+++ apps/nextjs/src/features/project/components/project-detail-modal/content/project-owner/input.tsx
@@ -3,7 +3,7 @@ import { memo, useCallback, useState } from 'react';
 import { Input as AtomsInput } from '@/components/ui/input';
 import { AssignProjectOwnerMenu } from '@/features/project/components/assign-project-owner-menu';
 import { useClickOutside } from '@/hooks/use-click-outside';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import { useProjectTeammatesCommand } from '@/store/entities/project-teammate';
 import type { Teammate } from '@/store/entities/teammate';
