+++ apps/nextjs/src/features/project/components/share-project-modal/body/share/send-form.tsx
@@ -6,7 +6,7 @@ import { Textarea } from '@/components/ui/textarea';
 import { Wrap, WrapItem } from '@/components/ui/wrap';
 import { InviteProjectTeammateMenu } from '@/features/teammate/components/invite-project-teammate-menu';
 import { InvitedTeammateChip } from '@/features/teammate/components/invited-teammate-chip';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import type { Teammate } from '@/store/entities/teammate';
 import { PermissionMenu } from '../permission-menu';
