+++ apps/nextjs/src/features/task-detail/components/task-detail-body/form/assignee/input.tsx
@@ -3,7 +3,7 @@ import { memo, useCallback, useState } from 'react';
 import { Input as AtomsInput } from '@/components/ui/input';
 import { AssigneeMenu } from '@/features/task/components/assignee-menu';
 import { useClickOutside } from '@/hooks/use-click-outside';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import { useTaskCommand } from '@/store/entities/task';
 import type { Teammate } from '@/store/entities/teammate';
