+++ apps/nextjs/src/features/task-detail/components/task-detail-body/form/projects/input.tsx
@@ -4,7 +4,7 @@ import { Flex } from '@/components/ui/flex';
 import { Input as AtomsInput, type InputProps } from '@/components/ui/input';
 import { ProjectMenu } from '@/features/project/components/project-menu';
 import { useClickOutside } from '@/hooks/use-click-outside';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import { useProjectTaskCommand } from '@/store/entities/project-task';
 
 type Props = {
