+++ apps/nextjs/src/features/task-detail/components/thumbnail-attachment/tooltip.tsx
@@ -4,7 +4,7 @@ import {
   Tooltip as MoleculesTooltip,
   type TooltipProps,
 } from '@/components/ui/tooltip';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import { useTaskFile } from '@/store/entities/task-file';
 import { useThumbnailAttachmentContext } from './context';
