+++ apps/nextjs/src/features/task/components/popover-assignee-input/popover-assignee-input.tsx
@@ -2,7 +2,7 @@ import React, { type PropsWithChildren, useCallback } from 'react';
 import { Link } from '@/components/ui/link';
 import { Popover } from '@/components/ui/popover';
 import { Portal } from '@/components/ui/portal';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import { Content } from './content';
 
 type Props = {
