+++ apps/nextjs/src/features/task/components/popover-due-date-picker/body.tsx
@@ -4,7 +4,7 @@ import { DatePicker } from '@/components/ui/date-picker';
 import { Flex, type FlexProps } from '@/components/ui/flex';
 import { Popover } from '@/components/ui/popover';
 import { Separator } from '@/components/ui/separator';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import { dateFns } from '@/lib/date-fns';
 import { DueTime } from './due-time';
