+++ apps/nextjs/src/features/task/components/popover-due-date-picker/popover-due-date-picker.tsx
@@ -3,7 +3,7 @@ import { useCallback } from 'react';
 import { Link } from '@/components/ui/link';
 import { Popover, type PopoverRootProps } from '@/components/ui/popover';
 import { Portal } from '@/components/ui/portal';
-import { type SystemStyleObject, useDisclosure } from '@/lib/chakra';
+import { type SystemStyleObject, useDisclosure } from '@/lib/chakra-ui';
 import { Body } from './body';
 
 type Props = {
