+++ apps/nextjs/src/features/task/components/tasks-board/tasks-board-list-item/more-action/more-action.tsx
@@ -3,7 +3,7 @@ import { Flex } from '@/components/ui/flex';
 import { Icon } from '@/components/ui/icon';
 import { IconButton } from '@/components/ui/icon-button';
 import { Menu } from '@/components/ui/menu';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import {
   useTasksBoardListItemContext,
   useTasksBoardListItemInputContext,
