+++ apps/nextjs/src/features/task/components/tasks-calendar/tasks-calendar-list-item/list-item/check-icon.tsx
@@ -2,7 +2,7 @@ import type React from 'react';
 import { memo, useCallback, useMemo } from 'react';
 import { CheckIcon as AtomsCheckIcon } from '@/components/ui/check-icon';
 import type { IconProps } from '@/components/ui/icon';
-import type { SystemStyleObject } from '@/lib/chakra';
+import type { SystemStyleObject } from '@/lib/chakra-ui';
 import { useProject } from '@/store/entities/project';
 import { useTask } from '@/store/entities/task';
