+++ apps/nextjs/src/features/task/components/tasks-list/tasks-list-cells/tasks-assignee/input.tsx
@@ -3,7 +3,7 @@ import { memo, useCallback, useState } from 'react';
 import { Input as AtomsInput } from '@/components/ui/input';
 import { AssignTaskMenu } from '@/features/task/components/assign-task-menu';
 import { useClickOutside } from '@/hooks/use-click-outside';
-import { useDisclosure } from '@/lib/chakra';
+import { useDisclosure } from '@/lib/chakra-ui';
 import type { Teammate } from '@/store/entities/teammate';
 
 type Props = {
