+++ apps/nextjs/src/hooks/styles/use-disabled-style.ts
@@ -1,5 +1,5 @@
 import { useMemo } from 'react';
-import type { SystemStyleObject } from '@/lib/chakra';
+import type { SystemStyleObject } from '@/lib/chakra-ui';
 
 type Props = SystemStyleObject;
